last fixes after rebase
